MODALES 04/02/2024 1:14 PM

+++ app/controllers/ciclo/D_ciclo.php
@@ -1,4 +1,5 @@
-<?php 
+<?php
+session_start();
 
 include('../../../config/conexion.php');
 
@@ -8,32 +9,40 @@ $id = $_GET['id'];
 mysqli_begin_transaction($cn);
 
 try {
-    // Eliminar relaciones en detalle_ciclo_turno
-    $sqlDeleteRelaciones = "DELETE FROM detalle_ciclo_turno WHERE id_ci = '$id'";
-    mysqli_query($cn, $sqlDeleteRelaciones);
+    // Obtener el nombre del ciclo antes de eliminarlo
+    $sql_select = "SELECT nombre_ci FROM ciclo WHERE id_ci = $id";
+    $result = mysqli_query($cn, $sql_select);
 
-    // Eliminar el registro en la tabla ciclo
-    $sqlDeleteCiclo = "DELETE FROM ciclo WHERE id_ci = '$id'";
-    mysqli_query($cn, $sqlDeleteCiclo);
+    if ($result && $row = mysqli_fetch_assoc($result)) {
+        $nombre_ciclo = $row['nombre_ci'];
 
-    // Confirmar la transacción
-    mysqli_commit($cn);
+        // Eliminar relaciones en detalle_ciclo_turno
+        $sqlDeleteRelaciones = "DELETE FROM detalle_ciclo_turno WHERE id_ci = '$id'";
+        mysqli_query($cn, $sqlDeleteRelaciones);
 
-    echo "Eliminación exitosa";
+        // Eliminar el registro en la tabla ciclo
+        $sqlDeleteCiclo = "DELETE FROM ciclo WHERE id_ci = '$id'";
+        mysqli_query($cn, $sqlDeleteCiclo);
 
-    header("Location: ../../../ciclo.php");
+        // Confirmar la transacción
+        mysqli_commit($cn);
 
-
-
+        $_SESSION['deleted_ciclo'] = "Ciclo eliminado: $nombre_ciclo";
+    } else {
+        $_SESSION['deleted_ciclo'] = "No se pudo obtener la información del ciclo con ID: $id";
+    }
 } catch (Exception $e) {
     // Revertir la transacción en caso de error
     mysqli_rollback($cn);
 
     // Manejar el error según tus necesidades
-    echo "Error en la eliminación: " . $e->getMessage();
+    $_SESSION['error_ciclo'] = "El ciclo $nombre_ciclo contiene matrículas activas.";
 }
 
 // Cerrar la conexión
 mysqli_close($cn);
-?>
 
+// Redirigir
+header("Location: ../../../ciclo.php");
+exit(); // Importante salir después de redirigir para evitar ejecución adicional
+?>
